backend/requirements.txt
pandas
numpy
scikit-learn 
onnxruntime-gpu
joblib